Remove path dependency on nrf52
Having that path dependency requires both repositories to be located in
the same directory. While that likely matches the setup of someone who
works on both crates, it will cause build errors for anyone attempting
to build nrf52-hal, who hasn't cloned nrf52 into the same directory
before.

For this reason, I believe path dependencies are inappropriate for
projects that are not located in the same repository (unless as a
temporary measure that is never merged into master).