Add basic SPIM API
I opted to replace the old code outright. To get SPI to work, I first
needed to understand it, and to really understand it, I needed to
implement it. Once I had done that, there wasn't much point in keeping
around the old code.

I've written this API to fit my current use case, talking to the DW1000.
I'm not sure if it fits other use cases. It certainly doesn't fit with
the trait in embedded-hal. I plan to investigate this at a later date,
but for now, this is a good start.