Add a crate for the adafruit nrf52 board
At this stage it is convenient to have some board support crates
in this same repo; it makes it easier to test and understand API
changes, and simpler for CI.

I don't know how many variants of the nrf52 boards are out there
and in common use, so there may well be a point at which we decide
to split off the boards directories.

All this crate does is define some nicer names for the pins and
includes a blinky example that I've flashed to my device.