add nrf52-hal-common crate
This commit takes what I believe to be the lowest maintenance approach
to sharing a common implementation between two device crates.

The common crate has a set of features that cause the underlying
svd2rust generated device crate to be imported as `target_device`.
The implementation in that crate then targets `target_device` rather
than directly using either `nrf52` or `nrf523840`.

This avoids littering the module with conditional code and/or macros.

The consuming hal crates re-export the items from the common crate
and augment with any other additional functionality.

One downside to this approach is that the `nrf-hal-common` crate cannot
be successfully compiled when multiple targets are built.  This is
fine in practice but does mean that building in a workspace will
fail when that workspace enables multiple devices.

For that reason this commit adjusts the travis config to build the
crates individually.