twim: remove chunking in `copy_write_then_read`
The previous implementation would send a TWI start condition for each chunk,
so the slave chip would see each as a separate transaction. This is not equivalent
to `write_then_read`, which was the original goal of adding chunking.

It seems TWIM has no way to send multiple buffers in a single transaction, therefore
the ability to chunk is removed.